requirements.txt
# Core dependencies
fastapi==0.118.2
uvicorn==0.37.0
sqlalchemy==2.0.43
pydantic==2.12.0
pydantic-core==2.41.1
pydantic-settings==2.11.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
jinja2==3.1.6
email-validator==2.3.0
psycopg2-binary==2.9.10
alembic==1.16.5
python-dotenv==1.1.1
starlette==0.48.0

# HTTP and API dependencies
httpx==0.28.1
requests==2.32.5
openai==2.2.0

# Cloud storage dependencies
google-cloud-storage==2.15.0
aiofiles==23.2.1
boto3==1.35.0

# NLP and text analysis dependencies
spacy==3.8.7
nltk==3.9.2
# After installing spacy, run: python -m spacy download en_core_web_sm

# Visualization dependencies
matplotlib==3.10.7
seaborn==0.13.2
pandas==2.3.3
pillow==11.3.0

# Utility dependencies
typing-extensions==4.15.0
anyio==4.11.0
click==8.3.0
